backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.3,<25.0
django-storages>=1.14,<2.0
Pillow>=10.4,<11.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
razorpay>=1.4.2
setuptools==69.5.1
